feat: add consts/log-macros modules
